管理者チェックを追加

+++ src/main/java/actions/EmployeeAction.java
@@ -43,6 +43,9 @@ public class EmployeeAction extends ActionBase {
      */
     public void index() throws ServletException, IOException {
 
+        //管理者かどうかのチェック追記
+        if(checkAdmin()) {//追記
+
         //指定されたページ数の一覧画面に表示するデータを取得
         int page = getPage();
         List<EmployeeView> employees = service.getPerPage(page);
@@ -64,6 +67,7 @@ public class EmployeeAction extends ActionBase {
 
         //一覧画面を表示
         forward(ForwardConst.FW_EMP_INDEX);
+        }//追記
     }
 
     /**
@@ -72,12 +76,14 @@ public class EmployeeAction extends ActionBase {
      * @throws IOException
      */
     public void entryNew() throws ServletException, IOException {
-
+        //管理者かどうかのチェック追記
+        if(checkAdmin()) {//追記
         putRequestScope(AttributeConst.TOKEN, getTokenId()); //CSRF対策用トークン
         putRequestScope(AttributeConst.EMPLOYEE, new EmployeeView()); //空の従業員インスタンス
 
         //新規登録画面を表示
         forward(ForwardConst.FW_EMP_NEW);
+        }
     }
 
 
@@ -89,7 +95,8 @@ public class EmployeeAction extends ActionBase {
     public void create() throws ServletException, IOException {
 
         //CSRF対策 tokenのチェック
-        if (checkToken()) {
+        //管理者かどうかのチェック追記
+        if(checkAdmin() && checkToken()) {
 
             //パラメータの値を元に従業員情報のインスタンスを作成する
             EmployeeView ev = new EmployeeView(
@@ -137,6 +144,10 @@ public class EmployeeAction extends ActionBase {
      * @throws IOException
      */
     public void show() throws ServletException, IOException {
+
+        //管理者かどうかのチェック追記
+        if(checkAdmin()) {//追記
+
         //idを条件に従業員データを取得
         EmployeeView ev = service.findOne(toNumber(getRequestParam(AttributeConst.EMP_ID)));
 
@@ -150,7 +161,8 @@ public class EmployeeAction extends ActionBase {
 
         //詳細画面を表示
         forward(ForwardConst.FW_EMP_SHOW);
-    }
+        }
+        }
 
     /**
      * 編集画面を表示する
@@ -159,6 +171,8 @@ public class EmployeeAction extends ActionBase {
      */
     public void edit() throws ServletException, IOException {
 
+        //管理者かどうかのチェック追記
+        if(checkAdmin()) {//追記
         //idを条件に従業員データを取得する
         EmployeeView ev = service.findOne(toNumber(getRequestParam(AttributeConst.EMP_ID)));
 
@@ -174,7 +188,7 @@ public class EmployeeAction extends ActionBase {
 
         //編集画面を表示する
         forward(ForwardConst.FW_EMP_EDIT);
-
+        }
     }
 
     /**
@@ -185,7 +199,8 @@ public class EmployeeAction extends ActionBase {
     public void update() throws ServletException, IOException {
 
         //CSRF対策 tokenのチェック
-        if (checkToken()) {
+        //管理者かどうかのチェック追記
+        if(checkAdmin() && checkToken()) {
             //パラメータの値を元に従業員情報のインスタンスを作成する
             EmployeeView ev = new EmployeeView(
                     toNumber(getRequestParam(AttributeConst.EMP_ID)),
@@ -232,7 +247,8 @@ public class EmployeeAction extends ActionBase {
     public void destroy() throws ServletException, IOException {
 
         //CSRF対策 tokenのチェック
-        if (checkToken()) {
+        //管理者かどうかのチェック追記
+        if(checkAdmin() && checkToken()) {
 
             //idを条件に従業員データを論理削除する
             service.destroy(toNumber(getRequestParam(AttributeConst.EMP_ID)));
@@ -244,4 +260,25 @@ public class EmployeeAction extends ActionBase {
             redirect(ForwardConst.ACT_EMP, ForwardConst.CMD_INDEX);
         }
     }
+
+    /**
+     * ログイン中の従業員が管理者かどうかチェックし、管理者でなければエラー画面を表示
+     * true: 管理者 false: 管理者ではない
+     * @throws ServletException
+     * @throws IOException
+     */
+
+   public boolean checkAdmin() throws ServletException, IOException {
+       //セッションからログイン中の従業員を取得
+       EmployeeView ev = (EmployeeView)getSessionScope(AttributeConst.LOGIN_EMP);
+
+       //管理者でなければエラー画面を表示
+       if(ev.getAdminFlag() != AttributeConst.ROLE_ADMIN.getIntegerValue()){
+
+           forward(ForwardConst.FW_ERR_UNKNOWN);
+           return false;
+       }else {
+           return true;
+       }
+   }
 }
